# openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/{id}": {
      "get": {
        "operationId": "UsersV1Controller_getPublic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public user profile without private account fields."
          }
        },
        "tags": [
          "v1-users"
        ]
      }
    },
    "/v1/me/profile": {
      "put": {
        "operationId": "UsersV1Controller_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "v1-users"
        ]
      }
    },
...
  },
  "info": {
    "title": "MDTBBS Public Client API",
    "description": "Stable V1 contract for web, official applications and third-party public clients. Use /api/v1/capabilities for feature discovery. External server integrations use the separate /api/external/v1 API key boundary. JSON V1 endpoints return the { data, meta } / { error, meta } envelope unless explicitly documented as a raw file or image response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "Same-origin MindFourm API"
    }
  ],
  "components": {
    "securitySchemes": {
      "MindAuthBearer": {
        "scheme": "bearer",
        "bearerFormat": "MindAuth access token",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}